Move timeline to island.is/web
