+++ apps/web/components/index.ts
@@ -7,7 +7,7 @@ export * from './SearchInput/SearchInput'
 export * from './Sticky/Sticky'
 export * from './Heading/Heading'
 export * from './LinkCardList/LinkCardList'
-export * from './Timeline/Timeline'
+export * from './Timeline/TimelineSection'
 export * from './SideMenu/SideMenu'
 export * from './EmailSignup/EmailSignup'
 export * from './StoryList/StoryList'
