+++ libs/island-ui/core/src/index.ts
@@ -1,6 +1,5 @@
 // Components
 export * from './lib/Pagination/Pagination'
-export * from './lib/Timeline/Timeline'
 export * from './lib/AsyncSearch/AsyncSearch'
 export * from './lib/Button/Button'
 export * from './lib/Logo/Logo'
